Add MEX Wrapper for METIS Graph Partitioning Library
Only uses

    METIS_PartGraphKway()

in the implementation, but we could conceivably add an option to
use

    METIS_PartGraphRecursive()

at some point in the future.  Supports the same options as the
existing MATLAB-implementation which invokes the 'gpmetis'
stand-alone executable through file communication.

We intend for this to complement and possibly replace the
existing 'partitionMetis' function.