Divergence Mappings: Switch to Vectorised Approach
This commit trades memory for speed using SORTROWS on a larger
array to guarantee the appropriate per-cell grouping behaviour
for the connecting cells and faces.  We also infer the local
position of the diagonal element by a counting procedure instead
of through FIND.

This brings nearly a factor 6.5 reduction of the runtime setting
up the mappings on a real field case with 372,588 active cells
and 1,269,007 inter-cell connections.

We nevertheless need to pay attention to the increased memory
requirements and my have to refine this approach if that becomes
a limiting resource for larger models.