Properly Discard Disabled Data in Record Parser
This commit ensures that we disregard everything that follows the
*first* unquoted '/' character that appears in a record-based
keyword.  We recently got a model that had the following data

    DATES
    2 'JAN' 1991 / 2 'JAN' 2010
    /

and this ended up properly confusing the record parser.

The approach here is to look for the terminating '/' character in
the unquoted substrings within the established record structure and
to discard all substrings following the first unquoted '/'.  As an
added benefit, this obviates the 'is_complete' helper function and
thereby removes one 'persistent' variable.