Fix MD5SUM For Character Vectors
The original implementation passed the return value of 'mxGetChars'
directly on to md5_append().  This does not work since mxGetChars
does not return a 'char*', but rather an 'mxChar*' instead.  Take
the more careful approach of copying the string data into a local
buffer before passing that buffer on to md5_append().

While here, also fix a format error which would occasionally output
a single character into the result string whereas each element of
the 'digest' must always correspond to exactly two characters in the
result string.  Finally, ensure that we release the temporary result
string acquired from mxCalloc to avoid memory leaks.