Don't Define MEX Flags If No C++ Compiler Available
We've had a number of user Forum posts that ultimately arise
from not having a C++ compiler installed on the local computer.
In this case the default diagnostic message generated by trying
to access fields in the (missing)

    mex.getCompilerConfigurations('c++')

structure does not provide enough information to allow the user
to conclude that they're missing a C++ compiler.

Add a check for this condition and emit a more helpful message
in order to preempt future posts of the same kind.