Simplify IsSamePartition Predicate
We don't have to look at representative cells.  If the two
partitions indeed are the same, just renumbered, then the number
of unique rows in [p1,p2] is the same as the number of coarse
blocks.

As an added benefit, this makes 'isSamePartition' agnostic to
the smallest block ID meaning

  isSamePartition(p, p - 1)

returns true even if min(p - 1) == 0.  This property is useful
when analysing partition vectors loaded from an external source
which might not create one-based block IDs.