Add Simple Alias Mechanism for Module Names
In particular, this enables giving a name for a collection of
modules--e.g., for backwards compatibility.  At the introduction
we add a single hard-coded alias for the original 'co2lab' module.
This is to cater to older scripts which have not yet been adapted to
the new split 'co2lab-*' modules.  The command

    mrstModule add co2lab

will then expand the name 'co2lab' to its constituent components.

At some point in the future we may consider exposing the alias
mechanism to users, but that will require more testing time than is
available just before the MRST 2024a release.