Add Fallback Option for Missing openedFiles Builtin
Function 'openedFiles' was introduced in MATLAB R2024a and made the
syntax

    fids = fopen('all')

obsolescent.  At the same time, FOPEN('all') started generating
warnings.

This commit makes a simple wrapper for 'openedFiles' that falls back
to FOPEN('all') if the new function is not available--e.g., when
running in older versions of MATLAB.